Fix offload failures: base-exception ratchet and daemon-requiring test placement
Two CI failures from the prior commits:

1. test_prevent_base_exception_catch ratchet: an autofix iteration added
   'except BaseException' in DockerProviderBackend.build_provider_instance to
   close the discarded instance on the empty/unavailable path. Replaced it with
   an explicit 'except MngrError' (the only thing has_state_container raises is
   ProviderUnavailableError) plus a dedicated close-before-raise on the
   ProviderEmptyError path -- same resource-leak safety, no BaseException catch.

2. test_read_only_construction_is_empty_and_creates_no_state_container ran in
   the offload sandboxes (no docker daemon) and failed on create_docker_client.
   It was in instance_test.py marked docker_sdk only, so the offload marker
   filter (which excludes acceptance/release, not docker_sdk) collected it.
   Moved it to test_docker_lifecycle.py (pytestmark=acceptance), matching the
   convention for real-daemon docker tests: acceptance keeps them out of
   offload and the test-docker job runs them with a daemon. Removed the
   now-unused imports from instance_test.py.
