+++ libs/mngr/imbue/mngr/providers/docker/backend.py
@@ -84,19 +84,25 @@ class DockerProviderBackend(ProviderBackendInterface):
             mngr_ctx=mngr_ctx,
             config=config,
         )
-        # The emptiness check materializes the instance's Docker client. If the
-        # provider declares itself empty (or the daemon is unreachable), the
-        # instance is discarded without being cached, so its close() would never
-        # run -- leaking the client connection. Close it here before re-raising.
+        if is_for_host_creation:
+            return instance
+        # Read-only construction must not bring the state container into
+        # existence. The emptiness check materializes the instance's Docker
+        # client, but the instance is discarded (never cached) when we raise, so
+        # its close() would never run -- leaking the client connection. Close it
+        # before propagating. has_state_container() raises ProviderUnavailableError
+        # (a MngrError) if the daemon is unreachable.
         try:
-            if not is_for_host_creation and not instance.has_state_container():
-                raise ProviderEmptyError(
-                    name,
-                    "no Docker state container exists yet (nothing has been created on this daemon)",
-                )
-        except BaseException:
+            state_container_exists = instance.has_state_container()
+        except MngrError:
             instance.close()
             raise
+        if not state_container_exists:
+            instance.close()
+            raise ProviderEmptyError(
+                name,
+                "no Docker state container exists yet (nothing has been created on this daemon)",
+            )
         return instance
