+++ libs/mngr/imbue/mngr/providers/docker/test_docker_lifecycle.py
@@ -7,6 +7,7 @@ import pytest
 from imbue.mngr.config.data_types import MngrContext
 from imbue.mngr.errors import HostNotFoundError
 from imbue.mngr.errors import MngrError
+from imbue.mngr.errors import ProviderEmptyError
 from imbue.mngr.errors import SnapshotNotFoundError
 from imbue.mngr.hosts.host import Host
 from imbue.mngr.hosts.offline_host import OfflineHost
@@ -14,11 +15,16 @@ from imbue.mngr.primitives import AgentId
 from imbue.mngr.primitives import HostId
 from imbue.mngr.primitives import HostName
 from imbue.mngr.primitives import ImageReference
+from imbue.mngr.primitives import ProviderInstanceName
 from imbue.mngr.primitives import SnapshotId
 from imbue.mngr.primitives import SnapshotName
+from imbue.mngr.providers.docker.backend import DockerProviderBackend
+from imbue.mngr.providers.docker.config import DockerProviderConfig
 from imbue.mngr.providers.docker.instance import DockerProviderInstance
+from imbue.mngr.providers.docker.instance import create_docker_client
 from imbue.mngr.providers.docker.testing import make_docker_provider
 from imbue.mngr.providers.docker.testing import make_docker_provider_with_cleanup
+from imbue.mngr.providers.docker.volume import state_container_name
 
 pytestmark = [pytest.mark.acceptance, pytest.mark.timeout(600)]
 
@@ -272,6 +278,42 @@ def test_close_closes_docker_client(temp_mngr_ctx: MngrContext) -> None:
     # Access the client to initialize it
     _ = provider._docker_client
     provider.close()
+
+
+@pytest.mark.docker_sdk
+def test_read_only_construction_is_empty_and_creates_no_state_container(temp_mngr_ctx: MngrContext) -> None:
+    """Building the docker provider for a read-only op must not create the state container.
+
+    Mirrors the Modal backend: when nothing has been created yet and
+    is_for_host_creation is False, build_provider_instance raises
+    ProviderEmptyError (so the provider loader skips docker) instead of lazily
+    materializing the singleton state container -- which is what caused
+    read-only commands like `mngr list` to leak state containers.
+    """
+    config = DockerProviderConfig(isolate_host_volumes=False)
+    user_id = str(temp_mngr_ctx.get_profile_user_id())
+    container_name = state_container_name(temp_mngr_ctx.config.prefix, user_id)
+
+    client = create_docker_client()
+    try:
+        with pytest.raises(ProviderEmptyError):
+            DockerProviderBackend.build_provider_instance(
+                name=ProviderInstanceName("docker"),
+                config=config,
+                mngr_ctx=temp_mngr_ctx,
+                is_for_host_creation=False,
+            )
+        # The read-only construction must not have created the state container.
+        with pytest.raises(docker.errors.NotFound):
+            client.containers.get(container_name)
+    finally:
+        # Defensive: if the assertion above regresses and a container WAS
+        # created, remove it so this test does not itself leak.
+        try:
+            client.containers.get(container_name).remove(force=True)
+        except docker.errors.NotFound:
+            pass
+        client.close()
 
 
 @pytest.mark.docker
